+++ app/Http/Controllers/StoreController.php
@@ -212,12 +212,13 @@ class StoreController extends Controller
         $owner = Owner::where('customer_id', $id)->first();
         // $this->deleteImage($product->image);
         $customer->deleted_by = Auth::user()->id;
-        $owner->deleted_by = Auth::user()->id;
+        if(!empty($owner)){
+            $owner->deleted_by = Auth::user()->id;
+            $owner->save();
+            $owner->delete();
+        }
         $customer->save();
-        $owner->save();
-
         $customer->delete();
-        $owner->delete();
 
         return response(['status' => 'success', 'message' => 'Toko berhasil dihapus']);
     }
@@ -248,7 +249,7 @@ class StoreController extends Controller
         $owner = Owner::onlyTrashed()->where('customer_id', $id)->first();
         // return $customer->photo;
         !empty($customer->photo) ? $this->deleteImage($customer->photo) : '';
-        $owner->forceDelete();
+        !empty($owner) ? $owner->forceDelete() : '';
         $customer->forceDelete();
 
         return response(['status' => 'success', 'message' => 'Toko berhasil dihapus permanen']);
